backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
langchain-community
langchain-text-splitters
python-dotenv
langchain-core
langchain-huggingface
SQLAlchemy
pypdf
faiss-cpu
PyJWT
Werkzeug
